Move export types to types.ts
